backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.1
sqlalchemy==2.0.22
psycopg2-binary==2.9.6
pydantic==1.10.12
pytest==7.4.2
httpx==0.24.0
